Code reorganisation
- Add `setup.py`, `requirements.txt`
- Move code to package `playground_bo`
- Move notebooks to `notebooks`
