+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    name="playground_bo",
+    packages=["playground_bo"],
+    scripts=[],
+)
